Convert spawns to use @Id annotation
