+++ game/src/main/kotlin/content/area/asgarnia/taverley/WitchesHouse.kt
@@ -2,6 +2,7 @@ package content.area.asgarnia.taverley
 
 import world.gregs.voidps.engine.Api
 import world.gregs.voidps.engine.data.definition.PatrolDefinitions
+import world.gregs.voidps.engine.entity.Id
 import world.gregs.voidps.engine.entity.character.mode.Patrol
 import world.gregs.voidps.engine.entity.character.npc.NPC
 import world.gregs.voidps.engine.event.Script
@@ -12,10 +13,9 @@ class WitchesHouse : Api {
 
     val patrols: PatrolDefinitions by inject()
 
+    @Id("nora_t_hagg")
     override fun spawn(npc: NPC) {
-        if (npc.id == "nora_t_hagg") {
-            val patrol = patrols.get("nora_t_hagg")
-            npc.mode = Patrol(npc, patrol.waypoints)
-        }
+        val patrol = patrols.get("nora_t_hagg")
+        npc.mode = Patrol(npc, patrol.waypoints)
     }
 }
