+++ game/src/main/kotlin/content/area/kandarin/ourania/ZamorakCrafter.kt
@@ -2,6 +2,7 @@ package content.area.kandarin.ourania
 
 import world.gregs.voidps.engine.Api
 import world.gregs.voidps.engine.data.definition.PatrolDefinitions
+import world.gregs.voidps.engine.entity.Id
 import world.gregs.voidps.engine.entity.character.mode.Patrol
 import world.gregs.voidps.engine.entity.character.npc.NPC
 import world.gregs.voidps.engine.entity.obj.GameObjects
@@ -17,11 +18,10 @@ class ZamorakCrafter : Api {
     val objects: GameObjects by inject()
     val patrols: PatrolDefinitions by inject()
 
+    @Id("zamorak_crafter*")
     override fun spawn(npc: NPC) {
-        if (npc.id.startsWith("zamorak_crafter")) {
-            val patrol = patrols.get(if (npc.id == "zamorak_crafter_start") "zamorak_crafter_to_altar" else "zamorak_crafter_to_bank")
-            npc.mode = Patrol(npc, patrol.waypoints)
-        }
+        val patrol = patrols.get(if (npc.id == "zamorak_crafter_start") "zamorak_crafter_to_altar" else "zamorak_crafter_to_bank")
+        npc.mode = Patrol(npc, patrol.waypoints)
     }
 
     override fun move(npc: NPC, from: Tile, to: Tile) {
